requirements.txt
Django==4.1
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
Pillow==9.2.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0